MSI-1306: Rename website breaks grid
- Switched from observer based logic to plugin based logic for website and eventual stock_link deletion
